refactor(board): extract useBoardCamera (D2 hook extraction, PR 2 of 3)
D2-board-tsx-decomposition-scoping.md calls this "the hard one" — the
manualCameraRef/manualCameraUntilRef gating, the fitCameraToContainerRef
mirror, and the ResizeObserver/double-rAF auto-fit effect all move verbatim
into client/src/components/board/useBoardCamera.ts, alongside the
board-reset effect and the tile-placed-auto-fit effect.

Sequencing note: the doc's own PR order was useBoardPointerControls (PR 3)
then useBoardCamera (PR 4). Implementing PR 3 first surfaced that its own
signature depends on this hook's output (camera setters + markManual), so
this PR reverses that order to match the real dependency direction.

A real ordering wrinkle, found and fixed rather than guessed past: the
camera *state* (useState) has to stay in Board.tsx, not move into this
hook — useBoardRenderLayout reads camera.scale for a debug trace, and that
value has to exist before this hook's fit logic (which needs `layout`,
useBoardRenderLayout's output) runs. useBoardCamera takes `setCamera` as a
parameter and owns the effects that decide the camera, not the state's
creation — preserving the original single-component hook-call order exactly.

Board.tsx: 673 -> 479 lines (now under the 500 max-lines budget).

Verification:
- tsc -b clean; full client vitest 230/230 files, 1726/1726 tests
- lint: 48/50 warnings (down from 49 — Board.tsx's max-lines warning is
  gone); lint:hooks (--max-warnings 0 gate) clean — two react-hooks
  warnings surfaced from refs/setters now crossing a hook boundary, fixed
  by adding the (stable) ref/setter identities to their dep arrays rather
  than suppressing
- board-camera.spec.ts (#176) loop-tested 10/10 clean against this
  extraction — the exact behaviour this hook is riskiest for
- fritz-play-to-completion.spec.ts (#175) run 3/3 clean as a broader
  full-match sanity check

Next: useBoardPointerControls (now genuinely last, consuming this hook's
camera setters) — its own PR, gated on board-camera.spec.ts staying green.
