+++ client/src/components/Board.tsx
@@ -3,7 +3,6 @@ import {
   forwardRef,
   memo,
   useCallback,
-  useEffect,
   useImperativeHandle,
   useRef,
   useState,
@@ -20,8 +19,8 @@ import {
   traceCameraDebug,
   traceDailyFritzBoardEvent,
 } from './boardDiagnostics';
-import { calculateBoardFitScale } from './board/boardLayout';
 import { useBoardRenderLayout } from './board/useBoardRenderLayout';
+import { useBoardCamera } from './board/useBoardCamera';
 
 // ─── Board Component ─────────────────────────────────────────
 
@@ -67,17 +66,6 @@ function highlightedEndsEqual(a?: number[] | null, b?: number[] | null): boolean
   return true;
 }
 
-function cameraStatesEqual(
-  a: { x: number; y: number; scale: number },
-  b: { x: number; y: number; scale: number },
-): boolean {
-  return (
-    Math.abs(a.x - b.x) < 0.5 &&
-    Math.abs(a.y - b.y) < 0.5 &&
-    Math.abs(a.scale - b.scale) < 0.001
-  );
-}
-
 function BoardComponent(
   {
     board,
@@ -109,12 +97,11 @@ function BoardComponent(
   if (profileDailyFritz) {
     recordDailyFritzBoardMetric('boardRenderCount', 1);
   }
-  const containerRef = useRef<HTMLDivElement>(null);
-  const fitRetryRafRef = useRef<number | null>(null);
-  const manualCameraRef = useRef(false);
-  const manualCameraUntilRef = useRef(0);
-  const lastResetSignatureRef = useRef('');
-  const [viewportSize, setViewportSize] = useState({ width: 0, height: 0 });
+  // Camera state lives here, not inside useBoardCamera below — its `.scale`
+  // needs to exist for useBoardRenderLayout's debug trace before `layout`
+  // (the value useBoardCamera's fit effects need) exists. Same render order
+  // as the original single-component code; useBoardCamera owns the effects
+  // that decide the camera, not the state itself.
   const [camera, setCamera] = useState({ x: 0, y: 0, scale: 1 });
   const [isDragging, setIsDragging] = useState(false);
   const dragStart = useRef({ x: 0, y: 0, camX: 0, camY: 0 });
@@ -134,215 +121,34 @@ function BoardComponent(
     cameraScale: camera.scale,
   });
 
-  useEffect(() => {
-    if (lastResetSignatureRef.current === resetSignature) return;
-    lastResetSignatureRef.current = resetSignature;
-    manualCameraRef.current = false;
-    traceCameraDebug('[camera-debug] setCamera', {
-      reason: 'board-reset',
-      x: 0,
-      y: 0,
-      scale: 1,
-    });
-    const resetCamera = { x: 0, y: 0, scale: 1 };
-    setCamera((prev) => (cameraStatesEqual(prev, resetCamera) ? prev : resetCamera));
-  }, [resetSignature]);
-
   // Convert layout units to pixels
   const unitToPixels = tileSize;
+
+  const {
+    containerRef,
+    viewportSize,
+    manualCameraRef,
+    markManualCamera,
+    fitCameraToContainer,
+    fitCameraToContainerRef,
+  } = useBoardCamera({
+    setCamera,
+    layout,
+    containFullBoard,
+    resolvedFitMode,
+    boardTileCount,
+    staticView,
+    unitToPixels,
+    minCameraScale,
+    staticFitMainline,
+    staticSpineAnchor,
+    resetSignature,
+  });
+
   // Calculate board center offset
   const centerX = (layout.minX + layout.maxX) / 2;
   const centerY =
     staticView && staticFitMainline ? 0 : (layout.minY + layout.maxY) / 2;
-  const markManualCamera = useCallback(() => {
-    const now = typeof performance !== 'undefined' && typeof performance.now === 'function' ? performance.now() : Date.now();
-    manualCameraRef.current = true;
-    manualCameraUntilRef.current = now + 1500;
-  }, []);
-
-  // Stable ref so the RAF retry callback always calls the latest version,
-  // avoiding a stale closure when layout/deps change between scheduling and firing.
-  const fitCameraToContainerRef = useRef<(reason: string, width?: number, height?: number, force?: boolean) => void>(() => {});
-
-  const fitCameraToContainer = useCallback((reason: string, width?: number, height?: number, force = false) => {
-    const container = containerRef.current;
-    if (!container) return;
-    if (manualCameraRef.current && !force) {
-      const now =
-        typeof performance !== 'undefined' && typeof performance.now === 'function' ? performance.now() : Date.now();
-      if (now < manualCameraUntilRef.current) return;
-      manualCameraRef.current = false;
-    }
-
-    const rect = container.getBoundingClientRect();
-    const containerWidth = width ?? rect.width;
-    const containerHeight = height ?? rect.height;
-    if (containerWidth < 2 || containerHeight < 2) {
-      // Mobile can report 0x0 before layout settles; retry next frame.
-      if (typeof window !== 'undefined' && fitRetryRafRef.current === null) {
-        fitRetryRafRef.current = window.requestAnimationFrame(() => {
-          fitRetryRafRef.current = null;
-          fitCameraToContainerRef.current('retry');
-        });
-      }
-      return;
-    }
-
-    // Calculate scale to fit
-    const layoutSpanUnits = Math.max(layout.maxX - layout.minX, layout.maxY - layout.minY);
-    const targetFill =
-      containFullBoard
-        ? 0.88
-        : resolvedFitMode === 'guided'
-        ? layoutSpanUnits <= 3
-          ? 1.08
-          : layoutSpanUnits <= 5
-            ? 0.94
-            : layoutSpanUnits <= 8
-              ? 0.88
-              : layoutSpanUnits >= 10
-                ? 0.93
-                : 0.9
-        : layoutSpanUnits <= 3
-          ? 0.52
-          : layoutSpanUnits <= 5
-            ? 0.65
-            : layoutSpanUnits <= 8
-              ? 0.82
-              : layoutSpanUnits >= 10
-                ? 0.92
-                : 0.9;
-    let maxFitScale =
-      containFullBoard
-        ? 1.8
-        : resolvedFitMode === 'guided'
-        ? boardTileCount <= 1
-          ? 4.4
-          : boardTileCount <= 4
-            ? 3.2
-            : boardTileCount <= 8
-              ? 2.2
-              : 1.6
-        : boardTileCount <= 1
-          ? 2.4
-          : boardTileCount <= 4
-            ? 2.1
-            : boardTileCount <= 8
-              ? 1.8
-              : 1.65;
-
-    if (staticView) {
-      const diagramFill = 1.04;
-      maxFitScale =
-        boardTileCount <= 3 ? 11.6 : boardTileCount <= 5 ? 10.4 : boardTileCount <= 8 ? 9.2 : 8;
-      const fitScale = calculateBoardFitScale({
-        layout,
-        tileSize: unitToPixels,
-        viewportWidth: containerWidth,
-        viewportHeight: containerHeight,
-        targetFill: diagramFill,
-        minScale: 0.22,
-        maxScale: maxFitScale,
-        multiplier: 2,
-      });
-      applyFitCamera(fitScale);
-      return;
-    }
-
-    const fitScale = calculateBoardFitScale({
-      layout,
-      tileSize: unitToPixels,
-      viewportWidth: containerWidth,
-      viewportHeight: containerHeight,
-      targetFill,
-      minScale: minCameraScale,
-      maxScale: maxFitScale,
-    });
-
-    applyFitCamera(fitScale);
-
-    function applyFitCamera(nextScale: number) {
-      let cameraY = 0;
-      if (staticView && staticFitMainline && staticSpineAnchor != null) {
-        const anchor = Math.min(0.85, Math.max(0.15, staticSpineAnchor));
-        cameraY = containerHeight * (anchor - 0.5);
-      }
-
-      traceCameraDebug('[camera-debug] setCamera', {
-        reason,
-        x: 0,
-        y: Number(cameraY.toFixed(1)),
-        scale: Number(nextScale.toFixed(3)),
-      });
-      const nextCamera = { x: 0, y: cameraY, scale: nextScale };
-      setCamera((prev) => (cameraStatesEqual(prev, nextCamera) ? prev : nextCamera));
-    }
-  }, [layout, containFullBoard, resolvedFitMode, boardTileCount, staticView, unitToPixels, minCameraScale, staticFitMainline, staticSpineAnchor, setCamera]);
-
-  useEffect(() => {
-    fitCameraToContainerRef.current = fitCameraToContainer;
-  }, [fitCameraToContainer]);
-
-  // Re-fit automatically when a new tile is placed, unless user has manually adjusted the camera.
-  const prevBoardTileCountRef = useRef(boardTileCount);
-  useEffect(() => {
-    if (staticView) return;
-    if (boardTileCount > prevBoardTileCountRef.current && !manualCameraRef.current) {
-      fitCameraToContainerRef.current('tile-placed-auto-fit');
-    }
-    prevBoardTileCountRef.current = boardTileCount;
-  }, [boardTileCount, staticView]);
-
-  // Single authoritative camera auto-fit: respond to layout and container size.
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-
-    const runFit = () => {
-      const rect = container.getBoundingClientRect();
-      setViewportSize((prev) =>
-        Math.abs(prev.width - rect.width) < 0.5 && Math.abs(prev.height - rect.height) < 0.5
-          ? prev
-          : { width: rect.width, height: rect.height },
-      );
-      fitCameraToContainer('effect-runFit', rect.width, rect.height);
-    };
-    runFit();
-    const raf1 = window.requestAnimationFrame(runFit);
-    const raf2 = window.requestAnimationFrame(() => window.requestAnimationFrame(runFit));
-
-    if (typeof ResizeObserver === 'undefined') {
-      return () => {
-        window.cancelAnimationFrame(raf1);
-        window.cancelAnimationFrame(raf2);
-      };
-    }
-
-    const observer = new ResizeObserver(() => {
-      runFit();
-    });
-    observer.observe(container);
-
-    return () => {
-      observer.disconnect();
-      window.cancelAnimationFrame(raf1);
-      window.cancelAnimationFrame(raf2);
-      if (fitRetryRafRef.current !== null) {
-        window.cancelAnimationFrame(fitRetryRafRef.current);
-        fitRetryRafRef.current = null;
-      }
-    };
-  }, [
-    fitCameraToContainer,
-    containFullBoard,
-    layout,
-    minCameraScale,
-    resolvedFitMode,
-    staticFitMainline,
-    staticSpineAnchor,
-    staticView,
-    unitToPixels,
-  ]);
 
   // Mouse wheel zoom
   const handleWheel = useCallback((e: React.WheelEvent) => {
@@ -415,7 +221,7 @@ function BoardComponent(
   const handleDoubleClick = useCallback(() => {
     manualCameraRef.current = false;
     fitCameraToContainer('double-click-reset', undefined, undefined, true);
-  }, [fitCameraToContainer]);
+  }, [fitCameraToContainer, manualCameraRef]);
 
   const applyZoomStep = useCallback((factor: number) => {
     markManualCamera();
@@ -437,7 +243,7 @@ function BoardComponent(
   const resetCameraToFit = useCallback(() => {
     manualCameraRef.current = false;
     fitCameraToContainerRef.current('manual-reset', undefined, undefined, true);
-  }, []);
+  }, [fitCameraToContainerRef, manualCameraRef]);
 
   useImperativeHandle(
     ref,
